Uploading GUI
This is as far as ive gotten with the GUI today. It works for basically every mark we need it to do. I added a feature where if a piece of data is selected and the "Feedback" column cant see the full text a textbox field will show the full text feedback. There is room for adding stuff with the GUI if needed.